refactor(examples): rename simple-chatbot to simple_chatbot and align schemas (#249)
