+++ examples/simple_chatbot/domain/dtos/chatbot_dto.py
