+++ examples/simple_chatbot/infrastructure/database/models/chatbot_model.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import DateTime, Integer, String, func
+from sqlalchemy import DateTime, Integer, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from src._core.infrastructure.persistence.rdb.database import Base
@@ -13,7 +13,7 @@ class ChatMessageModel(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     prompt: Mapped[str] = mapped_column(String(1000), nullable=False)
-    reply: Mapped[str] = mapped_column(String(4000), nullable=False)
+    reply: Mapped[str] = mapped_column(Text, nullable=False)
     tokens_used: Mapped[int] = mapped_column(Integer, nullable=False)
     created_at: Mapped[datetime] = mapped_column(
         DateTime, server_default=func.now(), nullable=False
