+++ examples/simple_chatbot/infrastructure/di/simple_chatbot_container.py
@@ -1,13 +1,13 @@
 from dependency_injector import containers, providers
-from src.simple_chatbot.domain.services.chatbot_service import ChatService
-from src.simple_chatbot.infrastructure.chatbot.pydantic_ai_chatbot import (
+
+from examples.simple_chatbot.domain.services.chatbot_service import ChatService
+from examples.simple_chatbot.infrastructure.chatbot.pydantic_ai_chatbot import (
     PydanticAIChatbot,
 )
-from src.simple_chatbot.infrastructure.chatbot.stub_chatbot import StubChatbot
-from src.simple_chatbot.infrastructure.repositories.chatbot_repository import (
+from examples.simple_chatbot.infrastructure.chatbot.stub_chatbot import StubChatbot
+from examples.simple_chatbot.infrastructure.repositories.chatbot_repository import (
     ChatbotRepository,
 )
-
 from src._core.config import settings
