+++ examples/simple_chatbot/infrastructure/repositories/chatbot_repository.py
@@ -1,11 +1,10 @@
-from src.simple_chatbot.domain.dtos.chatbot_dto import ChatMessageDTO
-from src.simple_chatbot.domain.protocols.chatbot_repository_protocol import (
+from examples.simple_chatbot.domain.dtos.chatbot_dto import ChatMessageDTO
+from examples.simple_chatbot.domain.protocols.chatbot_repository_protocol import (
     ChatbotRepositoryProtocol,
 )
-from src.simple_chatbot.infrastructure.database.models.chatbot_model import (
+from examples.simple_chatbot.infrastructure.database.models.chatbot_model import (
     ChatMessageModel,
 )
-
 from src._core.infrastructure.persistence.rdb.base_repository import BaseRepository
 from src._core.infrastructure.persistence.rdb.database import Database
