+++ tests/unit/simple_chatbot/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for the simple chatbot example."""
