+++ tests/unit/simple_chatbot/conftest.py
@@ -1,59 +1,12 @@
-import importlib.abc
-import importlib.machinery
-import sys
-from importlib.machinery import SourceFileLoader
-from pathlib import Path
-
-
-# ---------------------------------------------------------------------------
-# MetaPathFinder to redirect ``src.simple_chatbot`` to the example folder
-# so tests can be run inside ``tests/unit/simple_chatbot/`` directly.
-# ---------------------------------------------------------------------------
-class SimpleChatbotFinder(importlib.abc.MetaPathFinder):
-    def find_spec(self, fullname, path, target=None):
-        if fullname == "src.simple_chatbot" or fullname.startswith(
-            "src.simple_chatbot."
-        ):
-            parts = fullname.split(".")[2:]
-            base_dir = (
-                Path(__file__).parent.parent.parent.parent
-                / "examples"
-                / "simple-chatbot"
-            )
-            target_path = base_dir.joinpath(*parts)
-
-            if target_path.is_dir():
-                init_file = target_path / "__init__.py"
-                if init_file.exists():
-                    loader = SourceFileLoader(fullname, str(init_file))
-                    spec = importlib.machinery.ModuleSpec(
-                        fullname, loader, origin=str(init_file), is_package=True
-                    )
-                    spec.submodule_search_locations = [str(target_path)]
-                    return spec
-            else:
-                py_file = target_path.with_suffix(".py")
-                if py_file.exists():
-                    loader = SourceFileLoader(fullname, str(py_file))
-                    return importlib.machinery.ModuleSpec(
-                        fullname, loader, origin=str(py_file), is_package=False
-                    )
-        return None
-
-
-sys.meta_path.insert(0, SimpleChatbotFinder())
-
-
 import pytest
 import pytest_asyncio
 
 # Force registration of chatbot model onto Base.metadata
-from src.simple_chatbot.infrastructure.database.models.chatbot_model import (  # noqa: E402
+from examples.simple_chatbot.infrastructure.database.models.chatbot_model import (
     ChatMessageModel,
 )
-
-from src._core.infrastructure.persistence.rdb.config import DatabaseConfig  # noqa: E402
-from src._core.infrastructure.persistence.rdb.database import (  # noqa: E402
+from src._core.infrastructure.persistence.rdb.config import DatabaseConfig
+from src._core.infrastructure.persistence.rdb.database import (
     Base,
     Database,
 )
